Fixing issue #59 (Plotting PNGs the right way up).
In doing so we discovered that the 'catalog convention' (from image
analysis; the coordinate system one uses to view images) is flipped in
the y-direction from the 'database convention' (which records a user's
clicks).

The good way to resolve it would be to flip it at the database level.

This is not what was done!!!

What I have done, instead, is to flip the catalog into database
convention when manipulating images / doing the clustering, and then
outputting in catalog convention. So, when producing the images, I then
invert the y axis before saving the figure.